refactor(livestore): move storeOptions into StoreRegistry.ts
Co-locate storeOptions helper with RegistryStoreOptions since they are
tightly coupled - storeOptions exists solely to provide type inference
for RegistryStoreOptions.
