+++ packages/@livestore/livestore/src/mod.ts
@@ -41,7 +41,7 @@ export {
   createStore,
   createStorePromise,
 } from './store/create-store.ts'
-export { type RegistryStoreOptions, StoreRegistry } from './store/StoreRegistry.ts'
+export { type RegistryStoreOptions, StoreRegistry, storeOptions } from './store/StoreRegistry.ts'
 export { Store } from './store/store.ts'
 export {
   isLiveQueryDef,
@@ -60,6 +60,5 @@ export {
   type SubscribeOptions,
   type Unsubscribe,
 } from './store/store-types.ts'
-export { storeOptions } from './store/storeOptions.ts'
 export { exposeDebugUtils } from './utils/dev.ts'
 export * from './utils/stack-info.ts'
