+++ packages/@livestore/livestore/src/store/store.ts
@@ -50,10 +50,10 @@ import {
   type Queryable,
   type RefreshReason,
   type StoreCommitOptions,
+  type StoreConstructorParams,
   type StoreEventsOptions,
   type StoreInternals,
   StoreInternalsSymbol,
-  type StoreConstructorParams,
   type StoreOtel,
   type SubscribeOptions,
   type Unsubscribe,
